Fix ability to have multiple instances on the page
The previous implementation had issues with that. Only the last instance
was working.

I think the issue was that the event handler was passed instead of
defining a custom function for each element. Ah. Yes. It was because
the method redefined the handler but now the `self` reference was
different.