Defer external scripts
These were giving the other deferred scripts time to load and run before
paint because they were blocking the critical path. That should be
avoided in production.

It also messed with the "examples". You don't get the disappearing
content when you just block the critical path. That gives the other
scripts time to run and load on the second thread.