Make disclosure be class based
Typically, you would probably use some aria class... which I am not sure
why that is better than just use "hidden" attribute.

The point here is that it's unlikely that I the disclosure functionality
will be able to use the toggling of "hidden" on the maybe-script.

So right now, this shows an interesting issue. I can use it in the
example because the content remains hidden.