Allow src to be relative to the document
Before, I was using only the origin part of the document URL. That would
mean that building the absolute URL would not work if the src on the
custom element was defined as a document relative path.

E.g. `https://example.com/dogs/` and `src=./cats.js` would have looked for
`https://example.com/cats.js` instead of
`https://example.com/dogs/cats.js`.

Now, this should be fixed and the latter computed.