Use "loading status" instead of "state change"
This feels more accurate. While technically, it changes from loading to
loaded (success or failure) state, we don't really manage the loading
state. We assume that's the state it is in. We don't handle that.

When we get the status code of the loading response its the first time
we really care about the status of the script loading.

So that name feels a bit more accurate.